add SkipNavigationButton to its own file
