+++ static/app-strings.json
@@ -2059,5 +2059,8 @@
   "by %channelTitle%": "by %channelTitle%",
   "Reset": "Reset",
   "Reset to original (previous) publish date": "Reset to original (previous) publish date",
+  "Trending for #Art": "Trending for #Art",
+  "Trending for #Education": "Trending for #Education",
+  "Trending for #Technology": "Trending for #Technology",
   "--end--": "--end--"
 }
